Set up initial navigation for mobile app
